+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add mathpar environment
 
 ### Fixed
+* Do not override file icons from other plugins
 * Allow more characters in environment names
 
 ## [0.11.4] - 2025-11-01